Refactor project structure
