+++ ClashSharp/ClashSharp.slnx
@@ -2,8 +2,11 @@
   <Configurations>
     <Platform Name="x64" />
   </Configurations>
+  <Project Path="ClashSharp.Tests/ClashSharp.Tests.csproj">
+    <Platform Project="x64" />
+  </Project>
   <Project Path="ClashSharp/ClashSharp.csproj">
-    <Platform Solution="*|x64" Project="x64" />
+    <Platform Project="x64" />
     <Deploy />
   </Project>
 </Solution>
