+++ ClashSharp/ClashSharp/App.xaml
@@ -7,18 +7,8 @@
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
-                <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
+                <ResourceDictionary Source="Themes/Generic.xaml" />
             </ResourceDictionary.MergedDictionaries>
-
-            <Thickness x:Key="ClashPagePadding">32</Thickness>
-
-            <Style x:Key="ClashCardGridStyle" TargetType="Grid">
-                <Setter Property="Background" Value="{ThemeResource CardBackgroundFillColorDefaultBrush}" />
-                <Setter Property="BorderBrush" Value="{ThemeResource CardStrokeColorDefaultBrush}" />
-                <Setter Property="BorderThickness" Value="1" />
-                <Setter Property="CornerRadius" Value="4" />
-                <Setter Property="Padding" Value="16" />
-            </Style>
         </ResourceDictionary>
     </Application.Resources>
 </Application>
