--- ClashSharp/ClashSharp/Components/GuidedDialog.xaml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Window
-    x:Class="ClashSharp.Components.GuidedDialog"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:ClashSharp.Components"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Title="GuidedDialog">
-
-    <Window.SystemBackdrop>
-        <MicaBackdrop />
-    </Window.SystemBackdrop>
-
-    <Grid>
-
-    </Grid>
-</Window>
