+++ ClashSharp/ClashSharp/Components/SettingRow.xaml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<UserControl
+    x:Class="ClashSharp.Components.SettingRow"
+    x:Name="Root"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+    <Grid Style="{StaticResource ClashCardGridStyle}" MinHeight="60">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="Auto" />
+        </Grid.ColumnDefinitions>
+
+        <StackPanel VerticalAlignment="Center">
+            <TextBlock Text="{Binding Title, ElementName=Root}" Style="{ThemeResource BodyTextBlockStyle}" />
+            <TextBlock
+                Text="{Binding Description, ElementName=Root}"
+                Style="{ThemeResource CaptionTextBlockStyle}"
+                Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                TextWrapping="WrapWholeWords" />
+        </StackPanel>
+
+        <ContentPresenter
+            Grid.Column="1"
+            Margin="12,0,0,0"
+            VerticalAlignment="Center"
+            Content="{Binding ActionContent, ElementName=Root}" />
+    </Grid>
+</UserControl>
