+++ ClashSharp/ClashSharp/Service/CoreConfigurationService.cs
@@ -4,13 +4,11 @@
  *
  * @author: WaterRun
  * @file: Service/CoreConfigurationService.cs
- * @date: 2026-06-15
+ * @date: 2026-06-17
  */
 
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Globalization;
 using System.IO;
 using System.Text;
 using System.Threading;
@@ -70,7 +68,9 @@ public sealed class CoreConfigurationService
     /// <exception cref="ArgumentOutOfRangeException"><paramref name="mode"/> cannot be mapped to a mihomo mode.</exception>
     public CoreConfigurationState EnsureConfiguration(ClashSharpMode mode)
     {
-        return EnsureConfiguration(mode, ShouldEnableTransparentProxy(mode, AppSettingsService.Instance.TransparentProxyEnabled));
+        return EnsureConfiguration(
+            mode,
+            MihomoRuntimeConfigurationBuilder.ShouldEnableTransparentProxy(mode, AppSettingsService.Instance.TransparentProxyEnabled));
     }
 
     /// <summary>Ensures the local configuration directory and managed configuration file match <paramref name="mode"/> and transparent proxy preference.</summary>
@@ -126,8 +126,8 @@ public sealed class CoreConfigurationService
         }
 
         string normalizedProfileId = NormalizeProfileId(profileId);
-        string normalizedText = NormalizeConfigurationText(configurationText);
-        ValidateProfileShape(normalizedText);
+        string normalizedText = MihomoRuntimeConfigurationBuilder.NormalizeConfigurationText(configurationText);
+        MihomoProfileShapeValidator.Validate(normalizedText);
 
         string profileDirectory = GetProfileDirectoryPath(normalizedProfileId);
         string profileConfigPath = Path.Combine(profileDirectory, "config.yaml");
@@ -205,8 +205,9 @@ public sealed class CoreConfigurationService
             throw new FileNotFoundException("Imported profile configuration was not found.", profileConfigPath);
         }
 
-        string configurationText = NormalizeConfigurationText(await File.ReadAllTextAsync(profileConfigPath, cancellationToken).ConfigureAwait(false));
-        ValidateProfileShape(configurationText);
+        string configurationText = MihomoRuntimeConfigurationBuilder.NormalizeConfigurationText(
+            await File.ReadAllTextAsync(profileConfigPath, cancellationToken).ConfigureAwait(false));
+        MihomoProfileShapeValidator.Validate(configurationText);
         await ValidateWithBundledMihomoAsync(profileDirectory, profileConfigPath, cancellationToken).ConfigureAwait(false);
 
         return new ProfileImportResult(
@@ -216,63 +217,6 @@ public sealed class CoreConfigurationService
             MihomoProfileParserService.Instance.ParseNodes(configurationText).Count,
             MihomoProfileParserService.Instance.ParseRules(configurationText).Count,
             "配置校验通过。");
-    }
-
-    /// <summary>Builds a minimal safe mihomo configuration text for local startup probing.</summary>
-    /// <param name="mixedPort">Mixed HTTP and SOCKS proxy port in range [1, 65535].</param>
-    /// <param name="mode">Master takeover mode whose equivalent mihomo mode should be emitted.</param>
-    /// <returns>YAML configuration text with deterministic line endings.</returns>
-    /// <exception cref="ArgumentOutOfRangeException"><paramref name="mixedPort"/> is outside the valid TCP port range.</exception>
-    /// <exception cref="ArgumentOutOfRangeException"><paramref name="mode"/> cannot be mapped to a mihomo mode.</exception>
-    private static string BuildDefaultConfiguration(int mixedPort, ClashSharpMode mode, bool transparentProxyEnabled)
-    {
-        if (mixedPort is < 1 or > 65535)
-        {
-            throw new ArgumentOutOfRangeException(nameof(mixedPort), "Port must be in the range [1, 65535].");
-        }
-
-        string mihomoMode = MapToMihomoMode(mode);
-
-        List<string> lines =
-        [
-            "# Generated by Clash#.",
-            "# This minimal configuration is overwritten when a profile is selected.",
-            $"mixed-port: {mixedPort.ToString(CultureInfo.InvariantCulture)}",
-            "allow-lan: false",
-            $"mode: {mihomoMode}",
-            "log-level: info",
-            "external-controller: 127.0.0.1:9090",
-            "ipv6: true",
-        ];
-
-        if (transparentProxyEnabled)
-        {
-            AppendTunConfiguration(lines);
-        }
-
-        lines.AddRange(
-        [
-            "profile:",
-            "  store-selected: true",
-            "  store-fake-ip: true",
-            "proxies: []",
-            "proxy-groups:",
-            "  - name: GLOBAL",
-            "    type: select",
-            "    proxies:",
-            "      - DIRECT",
-            "  - name: DIRECT",
-            "    type: select",
-            "    proxies:",
-            "      - DIRECT",
-            "rules:",
-            "  - MATCH,DIRECT",
-            string.Empty,
-        ]);
-
-        return string.Join(
-            "\n",
-            lines);
     }
 
     /// <summary>Builds runtime configuration from the active imported profile when available, otherwise from the default profile.</summary>
@@ -286,166 +230,10 @@ public sealed class CoreConfigurationService
         if (File.Exists(profileConfigPath) && !StringComparer.Ordinal.Equals(activeProfileId, ProfileCatalogIds.BuiltInDirect))
         {
             string profileText = File.ReadAllText(profileConfigPath);
-            return OverrideRuntimeKeys(profileText, mixedPort, mode, transparentProxyEnabled);
+            return MihomoRuntimeConfigurationBuilder.OverrideRuntimeKeys(profileText, mixedPort, mode, transparentProxyEnabled);
         }
 
-        return BuildDefaultConfiguration(mixedPort, mode, transparentProxyEnabled);
-    }
-
-    /// <summary>Overrides Clash#-controlled runtime keys while preserving imported profile content.</summary>
-    /// <param name="configurationText">Imported configuration text. Must not be null.</param>
-    /// <param name="mixedPort">Mixed HTTP and SOCKS proxy port in range [1, 65535].</param>
-    /// <param name="mode">Master takeover mode whose equivalent mihomo mode should be emitted.</param>
-    /// <returns>Configuration text with updated runtime keys.</returns>
-    /// <exception cref="ArgumentNullException"><paramref name="configurationText"/> is null.</exception>
-    private static string OverrideRuntimeKeys(string configurationText, int mixedPort, ClashSharpMode mode, bool transparentProxyEnabled)
-    {
-        ArgumentNullException.ThrowIfNull(configurationText);
-
-        string[] lines = NormalizeConfigurationText(configurationText).Split('\n');
-        bool wroteMixedPort = false;
-        bool wroteMode = false;
-        string mihomoMode = MapToMihomoMode(mode);
-        StringBuilder builder = new();
-
-        for (int index = 0; index < lines.Length; index++)
-        {
-            string line = lines[index];
-            string trimmed = line.TrimStart();
-            if (line.Length == trimmed.Length && trimmed.StartsWith("mixed-port:", StringComparison.Ordinal))
-            {
-                builder.Append("mixed-port: ");
-                builder.AppendLine(mixedPort.ToString(CultureInfo.InvariantCulture));
-                wroteMixedPort = true;
-                continue;
-            }
-
-            if (line.Length == trimmed.Length && trimmed.StartsWith("mode:", StringComparison.Ordinal))
-            {
-                builder.Append("mode: ");
-                builder.AppendLine(mihomoMode);
-                wroteMode = true;
-                continue;
-            }
-
-            if (line.Length == trimmed.Length && trimmed.StartsWith("tun:", StringComparison.Ordinal))
-            {
-                while (index + 1 < lines.Length && CountLeadingSpaces(lines[index + 1]) > 0)
-                {
-                    index++;
-                }
-
-                continue;
-            }
-
-            builder.AppendLine(line);
-        }
-
-        if (!wroteMixedPort)
-        {
-            builder.Insert(0, $"mixed-port: {mixedPort.ToString(CultureInfo.InvariantCulture)}\n");
-        }
-
-        if (!wroteMode)
-        {
-            builder.Insert(0, $"mode: {mihomoMode}\n");
-        }
-
-        if (transparentProxyEnabled)
-        {
-            StringBuilder tunBuilder = new();
-            AppendTunConfiguration(tunBuilder);
-            builder.Insert(0, tunBuilder.ToString());
-        }
-
-        return builder.ToString();
-    }
-
-    /// <summary>Appends Clash#-controlled mihomo TUN configuration to a line list.</summary>
-    /// <param name="lines">Mutable configuration line list. Must not be null.</param>
-    private static void AppendTunConfiguration(List<string> lines)
-    {
-        ArgumentNullException.ThrowIfNull(lines);
-
-        lines.Add("tun:");
-        lines.Add("  enable: true");
-        lines.Add("  stack: system");
-        lines.Add("  auto-route: true");
-        lines.Add("  auto-detect-interface: true");
-        lines.Add("  strict-route: false");
-        lines.Add("  dns-hijack:");
-        lines.Add("    - any:53");
-    }
-
-    /// <summary>Appends Clash#-controlled mihomo TUN configuration to a string builder.</summary>
-    /// <param name="builder">Mutable configuration builder. Must not be null.</param>
-    private static void AppendTunConfiguration(StringBuilder builder)
-    {
-        ArgumentNullException.ThrowIfNull(builder);
-
-        builder.AppendLine("tun:");
-        builder.AppendLine("  enable: true");
-        builder.AppendLine("  stack: system");
-        builder.AppendLine("  auto-route: true");
-        builder.AppendLine("  auto-detect-interface: true");
-        builder.AppendLine("  strict-route: false");
-        builder.AppendLine("  dns-hijack:");
-        builder.AppendLine("    - any:53");
-    }
-
-    /// <summary>Maps a Clash# master takeover mode to a mihomo configuration mode.</summary>
-    /// <param name="mode">Master takeover mode to map.</param>
-    /// <returns>The mihomo mode string used in configuration YAML.</returns>
-    /// <exception cref="ArgumentOutOfRangeException"><paramref name="mode"/> cannot be mapped to a mihomo mode.</exception>
-    private static string MapToMihomoMode(ClashSharpMode mode)
-    {
-        return mode switch
-        {
-            ClashSharpMode.Disabled or ClashSharpMode.Standby => "direct",
-            ClashSharpMode.RuleTakeover => "rule",
-            ClashSharpMode.FullTakeover => "global",
-            _ => throw new ArgumentOutOfRangeException(nameof(mode), mode, "Unsupported Clash# configuration mode."),
-        };
-    }
-
-    /// <summary>Returns whether TUN should be enabled for a selected master mode.</summary>
-    /// <param name="mode">Selected master takeover mode.</param>
-    /// <param name="transparentProxyEnabled">User transparent proxy preference.</param>
-    /// <returns>True only for active takeover modes when the user preference is enabled.</returns>
-    private static bool ShouldEnableTransparentProxy(ClashSharpMode mode, bool transparentProxyEnabled)
-    {
-        return transparentProxyEnabled && mode is ClashSharpMode.RuleTakeover or ClashSharpMode.FullTakeover;
-    }
-
-    /// <summary>Normalizes configuration text line endings and ensures a trailing line break.</summary>
-    /// <param name="configurationText">Configuration text. Must not be null.</param>
-    /// <returns>Configuration text using LF line endings and one trailing line break.</returns>
-    /// <exception cref="ArgumentNullException"><paramref name="configurationText"/> is null.</exception>
-    private static string NormalizeConfigurationText(string configurationText)
-    {
-        ArgumentNullException.ThrowIfNull(configurationText);
-
-        string normalized = configurationText.Replace("\r\n", "\n", StringComparison.Ordinal).Replace('\r', '\n').Trim();
-        return normalized + "\n";
-    }
-
-    /// <summary>Validates that a profile contains the minimum mihomo sections required for import.</summary>
-    /// <param name="configurationText">Normalized configuration text. Must not be null.</param>
-    /// <exception cref="ArgumentNullException"><paramref name="configurationText"/> is null.</exception>
-    /// <exception cref="ArgumentException">The configuration does not look like a mihomo profile.</exception>
-    private static void ValidateProfileShape(string configurationText)
-    {
-        ArgumentNullException.ThrowIfNull(configurationText);
-
-        bool hasProxySource = ContainsTopLevelKey(configurationText, "proxies")
-            || ContainsTopLevelKey(configurationText, "proxy-providers");
-        bool hasProxyGroups = ContainsTopLevelKey(configurationText, "proxy-groups");
-        bool hasRules = ContainsTopLevelKey(configurationText, "rules");
-
-        if (!hasProxySource || !hasProxyGroups || !hasRules)
-        {
-            throw new ArgumentException("Downloaded configuration must contain proxies or proxy-providers, proxy-groups, and rules sections.", nameof(configurationText));
-        }
+        return MihomoRuntimeConfigurationBuilder.BuildDefaultConfiguration(mixedPort, mode, transparentProxyEnabled);
     }
 
     /// <summary>Runs bundled mihomo configuration validation when the binary is available.</summary>
@@ -533,48 +321,6 @@ public sealed class CoreConfigurationService
         {
             File.Delete(profileConfigPath);
         }
-    }
-
-    /// <summary>Returns whether <paramref name="configurationText"/> contains a top-level YAML key.</summary>
-    /// <param name="configurationText">Configuration text. Must not be null.</param>
-    /// <param name="key">Top-level key without the colon. Must not be null.</param>
-    /// <returns>True when the key is present at the start of a line; otherwise false.</returns>
-    private static bool ContainsTopLevelKey(string configurationText, string key)
-    {
-        ArgumentNullException.ThrowIfNull(configurationText);
-        ArgumentNullException.ThrowIfNull(key);
-
-        string prefix = key + ":";
-        foreach (string line in configurationText.Split('\n'))
-        {
-            if (line.StartsWith(prefix, StringComparison.Ordinal))
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /// <summary>Counts leading spaces in a YAML line.</summary>
-    /// <param name="line">Input line. Must not be null.</param>
-    /// <returns>Leading space count.</returns>
-    private static int CountLeadingSpaces(string line)
-    {
-        ArgumentNullException.ThrowIfNull(line);
-
-        int count = 0;
-        foreach (char character in line)
-        {
-            if (character != ' ')
-            {
-                break;
-            }
-
-            count++;
-        }
-
-        return count;
     }
 
     /// <summary>Normalizes a profile identifier so it is safe for a local directory name.</summary>
