+++ ClashSharp/ClashSharp/Service/MihomoRuntimeConfigurationBuilder.cs
@@ -1,0 +1,256 @@
+/*
+ * Mihomo Runtime Configuration Builder
+ * Builds deterministic mihomo runtime configuration fragments controlled by Clash#
+ *
+ * @author: WaterRun
+ * @file: Service/MihomoRuntimeConfigurationBuilder.cs
+ * @date: 2026-06-15
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Text;
+using ClashSharp.Model;
+
+namespace ClashSharp.Service;
+
+/// <summary>Builds deterministic mihomo runtime configuration text controlled by Clash#.</summary>
+/// <remarks>
+/// Invariants: Generated text uses LF line endings and Clash# controlled runtime keys are placed deterministically.
+/// Thread safety: Stateless methods are safe for concurrent calls.
+/// Side effects: None.
+/// </remarks>
+internal static class MihomoRuntimeConfigurationBuilder
+{
+    /// <summary>Builds a minimal safe mihomo configuration text for local startup probing.</summary>
+    /// <param name="mixedPort">Mixed HTTP and SOCKS proxy port in range [1, 65535].</param>
+    /// <param name="mode">Master takeover mode whose equivalent mihomo mode should be emitted.</param>
+    /// <param name="transparentProxyEnabled">True to include Clash# controlled TUN configuration.</param>
+    /// <returns>YAML configuration text with deterministic line endings.</returns>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="mixedPort"/> is outside the valid TCP port range.</exception>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="mode"/> cannot be mapped to a mihomo mode.</exception>
+    public static string BuildDefaultConfiguration(int mixedPort, ClashSharpMode mode, bool transparentProxyEnabled)
+    {
+        ValidateMixedPort(mixedPort);
+
+        string mihomoMode = MapToMihomoMode(mode);
+        List<string> lines =
+        [
+            "# Generated by Clash#.",
+            "# This minimal configuration is overwritten when a profile is selected.",
+            $"mixed-port: {mixedPort.ToString(CultureInfo.InvariantCulture)}",
+            "allow-lan: false",
+            $"mode: {mihomoMode}",
+            "log-level: info",
+            "external-controller: 127.0.0.1:9090",
+            "ipv6: true",
+        ];
+
+        if (transparentProxyEnabled)
+        {
+            AppendTunConfiguration(lines);
+        }
+
+        lines.AddRange(
+        [
+            "profile:",
+            "  store-selected: true",
+            "  store-fake-ip: true",
+            "proxies: []",
+            "proxy-groups:",
+            "  - name: GLOBAL",
+            "    type: select",
+            "    proxies:",
+            "      - DIRECT",
+            "  - name: DIRECT",
+            "    type: select",
+            "    proxies:",
+            "      - DIRECT",
+            "rules:",
+            "  - MATCH,DIRECT",
+            string.Empty,
+        ]);
+
+        return string.Join("\n", lines);
+    }
+
+    /// <summary>Overrides Clash#-controlled runtime keys while preserving imported profile content.</summary>
+    /// <param name="configurationText">Imported configuration text. Must not be null.</param>
+    /// <param name="mixedPort">Mixed HTTP and SOCKS proxy port in range [1, 65535].</param>
+    /// <param name="mode">Master takeover mode whose equivalent mihomo mode should be emitted.</param>
+    /// <param name="transparentProxyEnabled">True to replace any imported TUN section with Clash# controlled TUN configuration.</param>
+    /// <returns>Configuration text with updated runtime keys.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="configurationText"/> is null.</exception>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="mixedPort"/> is outside the valid TCP port range.</exception>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="mode"/> cannot be mapped to a mihomo mode.</exception>
+    public static string OverrideRuntimeKeys(string configurationText, int mixedPort, ClashSharpMode mode, bool transparentProxyEnabled)
+    {
+        ArgumentNullException.ThrowIfNull(configurationText);
+        ValidateMixedPort(mixedPort);
+
+        string[] lines = NormalizeConfigurationText(configurationText).Split('\n');
+        bool wroteMixedPort = false;
+        bool wroteMode = false;
+        string mihomoMode = MapToMihomoMode(mode);
+        StringBuilder builder = new();
+
+        for (int index = 0; index < lines.Length; index++)
+        {
+            string line = lines[index];
+            string trimmed = line.TrimStart();
+            if (line.Length == trimmed.Length && trimmed.StartsWith("mixed-port:", StringComparison.Ordinal))
+            {
+                builder.Append("mixed-port: ");
+                builder.Append(mixedPort.ToString(CultureInfo.InvariantCulture));
+                builder.Append('\n');
+                wroteMixedPort = true;
+                continue;
+            }
+
+            if (line.Length == trimmed.Length && trimmed.StartsWith("mode:", StringComparison.Ordinal))
+            {
+                builder.Append("mode: ");
+                builder.Append(mihomoMode);
+                builder.Append('\n');
+                wroteMode = true;
+                continue;
+            }
+
+            if (line.Length == trimmed.Length && trimmed.StartsWith("tun:", StringComparison.Ordinal))
+            {
+                while (index + 1 < lines.Length && CountLeadingSpaces(lines[index + 1]) > 0)
+                {
+                    index++;
+                }
+
+                continue;
+            }
+
+            builder.Append(line);
+            builder.Append('\n');
+        }
+
+        if (!wroteMixedPort)
+        {
+            builder.Insert(0, $"mixed-port: {mixedPort.ToString(CultureInfo.InvariantCulture)}\n");
+        }
+
+        if (!wroteMode)
+        {
+            builder.Insert(0, $"mode: {mihomoMode}\n");
+        }
+
+        if (transparentProxyEnabled)
+        {
+            StringBuilder tunBuilder = new();
+            AppendTunConfiguration(tunBuilder);
+            builder.Insert(0, tunBuilder.ToString());
+        }
+
+        return builder.ToString();
+    }
+
+    /// <summary>Returns whether TUN should be enabled for a selected master mode.</summary>
+    /// <param name="mode">Selected master takeover mode.</param>
+    /// <param name="transparentProxyEnabled">User transparent proxy preference.</param>
+    /// <returns>True only for active takeover modes when the user preference is enabled.</returns>
+    public static bool ShouldEnableTransparentProxy(ClashSharpMode mode, bool transparentProxyEnabled)
+    {
+        return transparentProxyEnabled && mode is ClashSharpMode.RuleTakeover or ClashSharpMode.FullTakeover;
+    }
+
+    /// <summary>Maps a Clash# master takeover mode to a mihomo configuration mode.</summary>
+    /// <param name="mode">Master takeover mode to map.</param>
+    /// <returns>The mihomo mode string used in configuration YAML.</returns>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="mode"/> cannot be mapped to a mihomo mode.</exception>
+    public static string MapToMihomoMode(ClashSharpMode mode)
+    {
+        return mode switch
+        {
+            ClashSharpMode.Disabled or ClashSharpMode.Standby => "direct",
+            ClashSharpMode.RuleTakeover => "rule",
+            ClashSharpMode.FullTakeover => "global",
+            _ => throw new ArgumentOutOfRangeException(nameof(mode), mode, "Unsupported Clash# configuration mode."),
+        };
+    }
+
+    /// <summary>Normalizes configuration text line endings and ensures a trailing line break.</summary>
+    /// <param name="configurationText">Configuration text. Must not be null.</param>
+    /// <returns>Configuration text using LF line endings and one trailing line break.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="configurationText"/> is null.</exception>
+    public static string NormalizeConfigurationText(string configurationText)
+    {
+        ArgumentNullException.ThrowIfNull(configurationText);
+
+        string normalized = configurationText.Replace("\r\n", "\n", StringComparison.Ordinal).Replace('\r', '\n').Trim();
+        return normalized + "\n";
+    }
+
+    /// <summary>Appends Clash#-controlled mihomo TUN configuration to a line list.</summary>
+    /// <param name="lines">Mutable configuration line list. Must not be null.</param>
+    /// <exception cref="ArgumentNullException"><paramref name="lines"/> is null.</exception>
+    private static void AppendTunConfiguration(List<string> lines)
+    {
+        ArgumentNullException.ThrowIfNull(lines);
+
+        lines.Add("tun:");
+        lines.Add("  enable: true");
+        lines.Add("  stack: system");
+        lines.Add("  auto-route: true");
+        lines.Add("  auto-detect-interface: true");
+        lines.Add("  strict-route: false");
+        lines.Add("  dns-hijack:");
+        lines.Add("    - any:53");
+    }
+
+    /// <summary>Appends Clash#-controlled mihomo TUN configuration to a string builder.</summary>
+    /// <param name="builder">Mutable configuration builder. Must not be null.</param>
+    /// <exception cref="ArgumentNullException"><paramref name="builder"/> is null.</exception>
+    private static void AppendTunConfiguration(StringBuilder builder)
+    {
+        ArgumentNullException.ThrowIfNull(builder);
+
+        builder.Append("tun:\n");
+        builder.Append("  enable: true\n");
+        builder.Append("  stack: system\n");
+        builder.Append("  auto-route: true\n");
+        builder.Append("  auto-detect-interface: true\n");
+        builder.Append("  strict-route: false\n");
+        builder.Append("  dns-hijack:\n");
+        builder.Append("    - any:53\n");
+    }
+
+    /// <summary>Counts leading spaces in a YAML line.</summary>
+    /// <param name="line">Input line. Must not be null.</param>
+    /// <returns>Leading space count.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="line"/> is null.</exception>
+    private static int CountLeadingSpaces(string line)
+    {
+        ArgumentNullException.ThrowIfNull(line);
+
+        int count = 0;
+        foreach (char character in line)
+        {
+            if (character != ' ')
+            {
+                break;
+            }
+
+            count++;
+        }
+
+        return count;
+    }
+
+    /// <summary>Validates a mixed proxy port.</summary>
+    /// <param name="mixedPort">Mixed HTTP and SOCKS proxy port in range [1, 65535].</param>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="mixedPort"/> is outside the valid TCP port range.</exception>
+    private static void ValidateMixedPort(int mixedPort)
+    {
+        if (mixedPort is < 1 or > 65535)
+        {
+            throw new ArgumentOutOfRangeException(nameof(mixedPort), "Port must be in the range [1, 65535].");
+        }
+    }
+}
