+++ ClashSharp/ClashSharp/Strings/LocalizationResources.cs
@@ -1,0 +1,792 @@
+/*
+ * Localization Resources
+ * Provides structured localized string dictionaries for every supported Clash# display language
+ *
+ * @author: WaterRun
+ * @file: Strings/LocalizationResources.cs
+ * @date: 2026-06-17
+ */
+
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using ClashSharp.Model;
+
+namespace ClashSharp.Strings;
+
+/// <summary>Provides immutable localized string dictionaries for every supported display language.</summary>
+/// <remarks>
+/// Invariants: Every supported language dictionary must contain the same key set as English.
+/// Thread safety: Immutable dictionaries are safe for concurrent reads after type initialization.
+/// Side effects: None.
+/// </remarks>
+internal static class LocalizationResources
+{
+    /// <summary>Gets immutable localized string dictionaries keyed by supported display language.</summary>
+    /// <value>Two-level immutable lookup table; never null and never mutated after type initialization.</value>
+    public static ReadOnlyDictionary<AppLanguage, ReadOnlyDictionary<string, string>> Translations { get; } = BuildTranslations();
+
+    private static ReadOnlyDictionary<AppLanguage, ReadOnlyDictionary<string, string>> BuildTranslations()
+    {
+        var table = new Dictionary<AppLanguage, ReadOnlyDictionary<string, string>>
+        {
+            [AppLanguage.SimplifiedChinese] = new(new Dictionary<string, string>
+            {
+                ["Nav.MasterControl"] = "主控",
+                ["Nav.Proxies"] = "代理",
+                ["Nav.ProxyNodes"] = "节点",
+                ["Nav.Profiles"] = "配置",
+                ["Nav.Links"] = "链接",
+                ["Nav.Connections"] = "连接",
+                ["Nav.Rules"] = "规则",
+                ["Nav.Statistics"] = "统计数据",
+                ["Nav.Logs"] = "日志",
+                ["Nav.Settings"] = "设置",
+                ["Page.Stub"] = "开发中…",
+                ["Page.MasterControl.Description"] = "控制 Clash# 的接管状态、核心运行和当前代理入口。",
+                ["Page.ProxyNodes.Description"] = "查看节点、地区、延迟和当前策略组选择。",
+                ["Page.Profiles.Description"] = "管理订阅、配置校验、自动更新和回滚。",
+                ["Page.Links.Description"] = "管理订阅链接、更新周期、检查状态和配置导入入口。",
+                ["Page.Connections.Description"] = "监视当前连接、进程、规则命中和代理路径。",
+                ["Page.Rules.Description"] = "查看规则集、匹配结果和接管策略。",
+                ["Page.Statistics.Description"] = "查看总量、配置、节点和时间维度的长期统计。",
+                ["Page.Logs.Description"] = "查看 SQLite 日志、占用空间和清理入口。",
+                ["Page.Settings.Description"] = "配置语言、透明代理、Windows 原生修复和中国大陆显示策略。",
+                ["Command.Add"] = "添加",
+                ["Command.Check"] = "检查",
+                ["Command.Import"] = "导入",
+                ["Command.PersistSnapshot"] = "写入快照",
+                ["Command.Refresh"] = "刷新",
+                ["Command.SetActive"] = "设为当前",
+                ["Command.TestLatency"] = "测试延迟",
+                ["Command.Update"] = "更新",
+                ["Command.Validate"] = "校验",
+                ["Command.Cleanup"] = "清理",
+                ["Command.Cancel"] = "取消",
+                ["Label.CurrentProfile"] = "当前配置",
+                ["Logs.Storage.Title"] = "本地记录占用",
+                ["Logs.StorageUsage.Format"] = "{0}，{1:N0} 条日志，{2:N0} 条连接记录",
+                ["Logs.Cleanup.Title"] = "清理本地记录",
+                ["Logs.Cleanup.Mode.ByDate"] = "按日期清理",
+                ["Logs.Cleanup.Mode.BySize"] = "按大小清理",
+                ["Logs.Cleanup.Mode.ByCount"] = "按条数清理",
+                ["Logs.Cleanup.Mode.All"] = "全部清理",
+                ["Logs.Cleanup.Parameter.KeepDays"] = "保留天数",
+                ["Logs.Cleanup.Parameter.TargetSize"] = "目标大小 MB",
+                ["Logs.Cleanup.Parameter.KeepLogCount"] = "保留日志条数",
+                ["Logs.Cleanup.Parameter.None"] = "无需参数",
+                ["Logs.Cleanup.Description.ByDate"] = "删除早于指定天数的日志、连接和流量快照。",
+                ["Logs.Cleanup.Description.BySize"] = "从最旧记录开始清理并压缩数据库，尽量降至目标大小以下。",
+                ["Logs.Cleanup.Description.ByCount"] = "仅限制日志表条数；连接、流量和规则统计保持不变。",
+                ["Logs.Cleanup.Description.All"] = "删除全部日志、连接、流量、节点健康和规则命中记录。",
+                ["Statistics.Total.Title"] = "总统计",
+                ["Statistics.Profile.Title"] = "配置统计",
+                ["Statistics.Node.Title"] = "节点统计",
+                ["Statistics.ByProfile.Title"] = "按配置",
+                ["Statistics.ByDate.Title"] = "按日期",
+                ["Statistics.ByNode.Title"] = "按节点",
+                ["Statistics.LogsShortcut.Title"] = "日志系统",
+                ["Statistics.LogsShortcut.Description"] = "查看 SQLite 日志、占用空间和清理入口。",
+                ["Statistics.OpenLogs"] = "打开日志",
+                ["Statistics.TotalTraffic.Format"] = "上传 {0} / 下载 {1}",
+                ["Statistics.ConnectionCount.Format"] = "{0:N0} 条连接记录",
+                ["Statistics.ProfileCount.Format"] = "{0:N0} 个配置聚合",
+                ["Statistics.SnapshotCount.Format"] = "{0:N0} 个时间快照",
+                ["Statistics.NodeCount.Format"] = "{0:N0} 个节点聚合 / {1:N0} 个健康记录",
+                ["Statistics.RuleCount.Format"] = "{0:N0} 条规则命中聚合",
+                ["Master.StatusControl.Title"] = "主状态控件",
+                ["Master.StatusControl.Description"] = "选择 Clash# 如何接管 Windows 网络。透明代理开关位于设置。",
+                ["Master.Mode.Disabled.Title"] = "未启用",
+                ["Master.Mode.Disabled.Description"] = "关闭核心并还原系统代理。",
+                ["Master.Mode.Standby.Title"] = "待命",
+                ["Master.Mode.Standby.Description"] = "核心运行，默认直连。",
+                ["Master.Mode.RuleTakeover.Title"] = "按规则接管",
+                ["Master.Mode.RuleTakeover.Description"] = "启动核心并使用规则策略接管网络。",
+                ["Master.Mode.FullTakeover.Title"] = "接管所有",
+                ["Master.Mode.FullTakeover.Description"] = "启动核心并使用全局策略接管网络。",
+                ["Master.Status.Core"] = "核心状态",
+                ["Master.Status.SystemProxy"] = "系统代理",
+                ["Master.Status.TransparentProxy"] = "透明代理",
+                ["Master.Status.CoreConfiguration"] = "核心配置",
+                ["Master.Status.ConfigurationUnavailable"] = "配置不可用",
+                ["Master.Status.CoreReady.Format"] = "就绪 · {0}",
+                ["Master.Status.CoreUnavailable"] = "核心不可用",
+                ["Master.Status.CoreStartFailed"] = "核心启动失败",
+                ["Master.Status.Running"] = "运行中",
+                ["Master.Status.NotRunning"] = "未运行",
+                ["Master.Status.On"] = "开启",
+                ["Master.Status.Off"] = "关闭",
+                ["Master.Status.Fallback"] = "已回退",
+                ["Master.Status.Unavailable"] = "不可用",
+                ["Master.Status.Standby"] = "待命",
+                ["Settings.Section.Language"] = "语言",
+                ["Settings.Language.Title"] = "界面语言",
+                ["Settings.Language.Description"] = "使用本地化资源显示主流语言界面。",
+                ["Settings.Section.Proxy"] = "代理",
+                ["Settings.TransparentProxy.Title"] = "透明代理",
+                ["Settings.TransparentProxy.Description"] = "启用 TUN 模式；失败时默认回退到系统代理。",
+                ["Settings.TunFallback.Title"] = "TUN 失败时回退",
+                ["Settings.TunFallback.Description"] = "透明代理启动失败时，自动改用 Windows 系统代理接管。",
+                ["Settings.MixedPort.Title"] = "混合端口",
+                ["Settings.MixedPort.Description"] = "系统代理和终端代理使用的本地 HTTP/SOCKS 入口。",
+                ["Settings.ConnectionSampling.Title"] = "后台连接采样",
+                ["Settings.ConnectionSampling.Description"] = "定期读取 mihomo 活动连接并写入 SQLite 统计。",
+                ["Settings.SamplingInterval.Title"] = "采样间隔",
+                ["Settings.SamplingInterval.Description"] = "后台连接采样的间隔，单位为秒。",
+                ["Settings.Section.WindowsNative"] = "Windows 原生",
+                ["Settings.WindowsNative.Title"] = "WSL、终端和 Microsoft Store 网络修正",
+                ["Settings.WindowsNative.Description"] = "每项单独诊断、单独应用，并提供可撤销动作。",
+                ["Settings.Wsl.Title"] = "WSL 网络修正",
+                ["Settings.Terminal.Title"] = "终端网络修正",
+                ["Settings.Store.Title"] = "Microsoft Store 网络修正",
+                ["Command.Diagnose"] = "诊断",
+                ["Command.Apply"] = "应用",
+                ["Command.Reset"] = "还原",
+                ["Diagnostic.NotRun"] = "未诊断",
+                ["Settings.CheckStaleProxy.Title"] = "启动时检查代理残留",
+                ["Settings.CheckStaleProxy.Description"] = "如果 Windows 代理仍指向 Clash# 本地端口，则按异常恢复策略处理。",
+                ["Settings.RestoreProxyOnExit.Title"] = "退出时还原系统代理",
+                ["Settings.RestoreProxyOnExit.Description"] = "关闭 Clash# 时停止核心并关闭 Windows 系统代理。",
+                ["Settings.ProxyRecoveryMode.Title"] = "异常恢复策略",
+                ["Settings.ProxyRecoveryMode.Description"] = "用于解决其它代理软件或 Clash# 异常关闭后残留系统代理的问题。",
+                ["Settings.ProxyRecoveryMode.Ignore"] = "不处理",
+                ["Settings.ProxyRecoveryMode.Enable"] = "恢复为代理开启",
+                ["Settings.ProxyRecoveryMode.Disable"] = "恢复为代理关闭",
+                ["Settings.Section.MainlandChina"] = "中国大陆",
+                ["Settings.MainlandChinaDisplay.Title"] = "适用于中国大陆的文本和旗帜替换",
+                ["Settings.MainlandChinaDisplay.Description"] = "默认开启；敏感词过滤仅影响界面显示，不影响复制、搜索和导出。",
+            }),
+            [AppLanguage.TraditionalChinese] = new(new Dictionary<string, string>
+            {
+                ["Nav.MasterControl"] = "主控",
+                ["Nav.Proxies"] = "代理",
+                ["Nav.ProxyNodes"] = "節點",
+                ["Nav.Profiles"] = "設定檔",
+                ["Nav.Links"] = "連結",
+                ["Nav.Connections"] = "連線",
+                ["Nav.Rules"] = "規則",
+                ["Nav.Statistics"] = "統計資料",
+                ["Nav.Logs"] = "日誌",
+                ["Nav.Settings"] = "設定",
+                ["Page.Stub"] = "開發中…",
+                ["Page.MasterControl.Description"] = "控制 Clash# 的接管狀態、核心執行和目前代理入口。",
+                ["Page.ProxyNodes.Description"] = "檢視節點、地區、延遲和目前策略組選擇。",
+                ["Page.Profiles.Description"] = "管理訂閱、設定檔校驗、自動更新和回復。",
+                ["Page.Links.Description"] = "管理訂閱連結、更新週期、檢查狀態和設定檔匯入入口。",
+                ["Page.Connections.Description"] = "監視目前連線、處理程序、規則命中和代理路徑。",
+                ["Page.Rules.Description"] = "檢視規則集、匹配結果和接管策略。",
+                ["Page.Statistics.Description"] = "檢視總量、設定檔、節點和時間維度的長期統計。",
+                ["Page.Logs.Description"] = "檢視 SQLite 日誌、占用空間和清理入口。",
+                ["Page.Settings.Description"] = "設定語言、透明代理、Windows 原生修復和中國大陸顯示策略。",
+                ["Command.Add"] = "新增",
+                ["Command.Check"] = "檢查",
+                ["Command.Import"] = "匯入",
+                ["Command.PersistSnapshot"] = "寫入快照",
+                ["Command.Refresh"] = "重新整理",
+                ["Command.SetActive"] = "設為目前",
+                ["Command.TestLatency"] = "測試延遲",
+                ["Command.Update"] = "更新",
+                ["Command.Validate"] = "校驗",
+                ["Command.Cleanup"] = "清理",
+                ["Command.Cancel"] = "取消",
+                ["Label.CurrentProfile"] = "目前設定檔",
+                ["Logs.Storage.Title"] = "本機記錄占用",
+                ["Logs.StorageUsage.Format"] = "{0}，{1:N0} 筆日誌，{2:N0} 筆連線記錄",
+                ["Logs.Cleanup.Title"] = "清理本機記錄",
+                ["Logs.Cleanup.Mode.ByDate"] = "依日期清理",
+                ["Logs.Cleanup.Mode.BySize"] = "依大小清理",
+                ["Logs.Cleanup.Mode.ByCount"] = "依筆數清理",
+                ["Logs.Cleanup.Mode.All"] = "全部清理",
+                ["Logs.Cleanup.Parameter.KeepDays"] = "保留天數",
+                ["Logs.Cleanup.Parameter.TargetSize"] = "目標大小 MB",
+                ["Logs.Cleanup.Parameter.KeepLogCount"] = "保留日誌筆數",
+                ["Logs.Cleanup.Parameter.None"] = "無需參數",
+                ["Logs.Cleanup.Description.ByDate"] = "刪除早於指定天數的日誌、連線和流量快照。",
+                ["Logs.Cleanup.Description.BySize"] = "從最舊記錄開始清理並壓縮資料庫，盡量降至目標大小以下。",
+                ["Logs.Cleanup.Description.ByCount"] = "僅限制日誌表筆數；連線、流量和規則統計保持不變。",
+                ["Logs.Cleanup.Description.All"] = "刪除全部日誌、連線、流量、節點健康和規則命中記錄。",
+                ["Statistics.Total.Title"] = "總統計",
+                ["Statistics.Profile.Title"] = "設定檔統計",
+                ["Statistics.Node.Title"] = "節點統計",
+                ["Statistics.ByProfile.Title"] = "依設定檔",
+                ["Statistics.ByDate.Title"] = "依日期",
+                ["Statistics.ByNode.Title"] = "依節點",
+                ["Statistics.LogsShortcut.Title"] = "日誌系統",
+                ["Statistics.LogsShortcut.Description"] = "檢視 SQLite 日誌、占用空間和清理入口。",
+                ["Statistics.OpenLogs"] = "開啟日誌",
+                ["Statistics.TotalTraffic.Format"] = "上傳 {0} / 下載 {1}",
+                ["Statistics.ConnectionCount.Format"] = "{0:N0} 筆連線記錄",
+                ["Statistics.ProfileCount.Format"] = "{0:N0} 個設定檔聚合",
+                ["Statistics.SnapshotCount.Format"] = "{0:N0} 個時間快照",
+                ["Statistics.NodeCount.Format"] = "{0:N0} 個節點聚合 / {1:N0} 個健康記錄",
+                ["Statistics.RuleCount.Format"] = "{0:N0} 條規則命中聚合",
+                ["Master.StatusControl.Title"] = "主狀態控制",
+                ["Master.StatusControl.Description"] = "選擇 Clash# 如何接管 Windows 網路。透明代理開關位於設定。",
+                ["Master.Mode.Disabled.Title"] = "未啟用",
+                ["Master.Mode.Disabled.Description"] = "關閉核心並還原系統代理。",
+                ["Master.Mode.Standby.Title"] = "待命",
+                ["Master.Mode.Standby.Description"] = "核心執行，預設直連。",
+                ["Master.Mode.RuleTakeover.Title"] = "依規則接管",
+                ["Master.Mode.RuleTakeover.Description"] = "啟動核心並使用規則策略接管網路。",
+                ["Master.Mode.FullTakeover.Title"] = "接管全部",
+                ["Master.Mode.FullTakeover.Description"] = "啟動核心並使用全域策略接管網路。",
+                ["Master.Status.Core"] = "核心狀態",
+                ["Master.Status.SystemProxy"] = "系統代理",
+                ["Master.Status.TransparentProxy"] = "透明代理",
+                ["Master.Status.CoreConfiguration"] = "核心設定",
+                ["Master.Status.ConfigurationUnavailable"] = "設定不可用",
+                ["Master.Status.CoreReady.Format"] = "就緒 · {0}",
+                ["Master.Status.CoreUnavailable"] = "核心不可用",
+                ["Master.Status.CoreStartFailed"] = "核心啟動失敗",
+                ["Master.Status.Running"] = "執行中",
+                ["Master.Status.NotRunning"] = "未執行",
+                ["Master.Status.On"] = "開啟",
+                ["Master.Status.Off"] = "關閉",
+                ["Master.Status.Fallback"] = "已回退",
+                ["Master.Status.Unavailable"] = "不可用",
+                ["Master.Status.Standby"] = "待命",
+                ["Settings.Section.Language"] = "語言",
+                ["Settings.Language.Title"] = "介面語言",
+                ["Settings.Language.Description"] = "使用本地化資源顯示主流語言介面。",
+                ["Settings.Section.Proxy"] = "代理",
+                ["Settings.TransparentProxy.Title"] = "透明代理",
+                ["Settings.TransparentProxy.Description"] = "啟用 TUN 模式；失敗時預設回退到系統代理。",
+                ["Settings.TunFallback.Title"] = "TUN 失敗時回退",
+                ["Settings.TunFallback.Description"] = "透明代理啟動失敗時，自動改用 Windows 系統代理接管。",
+                ["Settings.MixedPort.Title"] = "混合連接埠",
+                ["Settings.MixedPort.Description"] = "系統代理和終端代理使用的本機 HTTP/SOCKS 入口。",
+                ["Settings.ConnectionSampling.Title"] = "背景連線採樣",
+                ["Settings.ConnectionSampling.Description"] = "定期讀取 mihomo 目前連線並寫入 SQLite 統計。",
+                ["Settings.SamplingInterval.Title"] = "採樣間隔",
+                ["Settings.SamplingInterval.Description"] = "背景連線採樣的間隔，單位為秒。",
+                ["Settings.Section.WindowsNative"] = "Windows 原生",
+                ["Settings.WindowsNative.Title"] = "WSL、終端機和 Microsoft Store 網路修正",
+                ["Settings.WindowsNative.Description"] = "每項單獨診斷、單獨套用，並提供可復原動作。",
+                ["Settings.Wsl.Title"] = "WSL 網路修正",
+                ["Settings.Terminal.Title"] = "終端機網路修正",
+                ["Settings.Store.Title"] = "Microsoft Store 網路修正",
+                ["Command.Diagnose"] = "診斷",
+                ["Command.Apply"] = "套用",
+                ["Command.Reset"] = "還原",
+                ["Diagnostic.NotRun"] = "未診斷",
+                ["Settings.CheckStaleProxy.Title"] = "啟動時檢查代理殘留",
+                ["Settings.CheckStaleProxy.Description"] = "如果 Windows 代理仍指向 Clash# 本機連接埠，則依異常復原策略處理。",
+                ["Settings.RestoreProxyOnExit.Title"] = "結束時還原系統代理",
+                ["Settings.RestoreProxyOnExit.Description"] = "關閉 Clash# 時停止核心並關閉 Windows 系統代理。",
+                ["Settings.ProxyRecoveryMode.Title"] = "異常復原策略",
+                ["Settings.ProxyRecoveryMode.Description"] = "用於解決其它代理軟體或 Clash# 異常關閉後殘留系統代理的問題。",
+                ["Settings.ProxyRecoveryMode.Ignore"] = "不處理",
+                ["Settings.ProxyRecoveryMode.Enable"] = "復原為代理開啟",
+                ["Settings.ProxyRecoveryMode.Disable"] = "復原為代理關閉",
+                ["Settings.Section.MainlandChina"] = "中國大陸",
+                ["Settings.MainlandChinaDisplay.Title"] = "適用於中國大陸的文字和旗幟替換",
+                ["Settings.MainlandChinaDisplay.Description"] = "預設開啟；敏感詞過濾僅影響介面顯示，不影響複製、搜尋和匯出。",
+            }),
+            [AppLanguage.English] = new(new Dictionary<string, string>
+            {
+                ["Nav.MasterControl"] = "Master Control",
+                ["Nav.Proxies"] = "Proxies",
+                ["Nav.ProxyNodes"] = "Nodes",
+                ["Nav.Profiles"] = "Profiles",
+                ["Nav.Links"] = "Links",
+                ["Nav.Connections"] = "Connections",
+                ["Nav.Rules"] = "Rules",
+                ["Nav.Statistics"] = "Statistics",
+                ["Nav.Logs"] = "Logs",
+                ["Nav.Settings"] = "Settings",
+                ["Page.Stub"] = "Under development…",
+                ["Page.MasterControl.Description"] = "Control Clash# takeover state, core runtime, and current proxy entry.",
+                ["Page.ProxyNodes.Description"] = "Review nodes, regions, latency, and active proxy group choices.",
+                ["Page.Profiles.Description"] = "Manage subscriptions, configuration validation, automatic updates, and rollback.",
+                ["Page.Links.Description"] = "Manage subscription links, update intervals, check status, and profile import entry points.",
+                ["Page.Connections.Description"] = "Monitor active connections, processes, rule hits, and proxy routes.",
+                ["Page.Rules.Description"] = "Inspect rule providers, match results, and takeover policy.",
+                ["Page.Statistics.Description"] = "Review long-term totals by profile, node, and time range.",
+                ["Page.Logs.Description"] = "Review SQLite logs, storage usage, and cleanup actions.",
+                ["Page.Settings.Description"] = "Configure language, transparent proxy, native Windows repair, and mainland China display policy.",
+                ["Command.Add"] = "Add",
+                ["Command.Check"] = "Check",
+                ["Command.Import"] = "Import",
+                ["Command.PersistSnapshot"] = "Save snapshot",
+                ["Command.Refresh"] = "Refresh",
+                ["Command.SetActive"] = "Set active",
+                ["Command.TestLatency"] = "Test latency",
+                ["Command.Update"] = "Update",
+                ["Command.Validate"] = "Validate",
+                ["Command.Cleanup"] = "Cleanup",
+                ["Command.Cancel"] = "Cancel",
+                ["Label.CurrentProfile"] = "Current profile",
+                ["Logs.Storage.Title"] = "Local storage usage",
+                ["Logs.StorageUsage.Format"] = "{0}, {1:N0} logs, {2:N0} connection records",
+                ["Logs.Cleanup.Title"] = "Clean local records",
+                ["Logs.Cleanup.Mode.ByDate"] = "Clean by date",
+                ["Logs.Cleanup.Mode.BySize"] = "Clean by size",
+                ["Logs.Cleanup.Mode.ByCount"] = "Clean by count",
+                ["Logs.Cleanup.Mode.All"] = "Clean all",
+                ["Logs.Cleanup.Parameter.KeepDays"] = "Days to keep",
+                ["Logs.Cleanup.Parameter.TargetSize"] = "Target size MB",
+                ["Logs.Cleanup.Parameter.KeepLogCount"] = "Logs to keep",
+                ["Logs.Cleanup.Parameter.None"] = "No parameter",
+                ["Logs.Cleanup.Description.ByDate"] = "Delete logs, connections, and traffic snapshots older than the selected number of days.",
+                ["Logs.Cleanup.Description.BySize"] = "Delete oldest records and compact the database until it is near the target size.",
+                ["Logs.Cleanup.Description.ByCount"] = "Only limit the Logs table count; connections, traffic, and rule statistics are preserved.",
+                ["Logs.Cleanup.Description.All"] = "Delete all logs, connections, traffic, node health, and rule-hit records.",
+                ["Statistics.Total.Title"] = "Totals",
+                ["Statistics.Profile.Title"] = "Profiles",
+                ["Statistics.Node.Title"] = "Nodes",
+                ["Statistics.ByProfile.Title"] = "By profile",
+                ["Statistics.ByDate.Title"] = "By date",
+                ["Statistics.ByNode.Title"] = "By node",
+                ["Statistics.LogsShortcut.Title"] = "Log system",
+                ["Statistics.LogsShortcut.Description"] = "Review SQLite logs, storage usage, and cleanup actions.",
+                ["Statistics.OpenLogs"] = "Open logs",
+                ["Statistics.TotalTraffic.Format"] = "Upload {0} / download {1}",
+                ["Statistics.ConnectionCount.Format"] = "{0:N0} connection records",
+                ["Statistics.ProfileCount.Format"] = "{0:N0} profile aggregates",
+                ["Statistics.SnapshotCount.Format"] = "{0:N0} time snapshots",
+                ["Statistics.NodeCount.Format"] = "{0:N0} node aggregates / {1:N0} health records",
+                ["Statistics.RuleCount.Format"] = "{0:N0} rule-hit aggregates",
+                ["Master.StatusControl.Title"] = "Primary status control",
+                ["Master.StatusControl.Description"] = "Choose how Clash# takes over Windows networking. Transparent proxy is configured in Settings.",
+                ["Master.Mode.Disabled.Title"] = "Disabled",
+                ["Master.Mode.Disabled.Description"] = "Stop the core and restore the system proxy.",
+                ["Master.Mode.Standby.Title"] = "Standby",
+                ["Master.Mode.Standby.Description"] = "Run the core with direct connection by default.",
+                ["Master.Mode.RuleTakeover.Title"] = "Rule takeover",
+                ["Master.Mode.RuleTakeover.Description"] = "Run the core and take over networking with rule policy.",
+                ["Master.Mode.FullTakeover.Title"] = "Full takeover",
+                ["Master.Mode.FullTakeover.Description"] = "Run the core and take over networking with global policy.",
+                ["Master.Status.Core"] = "Core status",
+                ["Master.Status.SystemProxy"] = "System proxy",
+                ["Master.Status.TransparentProxy"] = "Transparent proxy",
+                ["Master.Status.CoreConfiguration"] = "Core configuration",
+                ["Master.Status.ConfigurationUnavailable"] = "Configuration unavailable",
+                ["Master.Status.CoreReady.Format"] = "Ready · {0}",
+                ["Master.Status.CoreUnavailable"] = "Core unavailable",
+                ["Master.Status.CoreStartFailed"] = "Core startup failed",
+                ["Master.Status.Running"] = "Running",
+                ["Master.Status.NotRunning"] = "Not running",
+                ["Master.Status.On"] = "On",
+                ["Master.Status.Off"] = "Off",
+                ["Master.Status.Fallback"] = "Fallback",
+                ["Master.Status.Unavailable"] = "Unavailable",
+                ["Master.Status.Standby"] = "Standby",
+                ["Settings.Section.Language"] = "Language",
+                ["Settings.Language.Title"] = "Display language",
+                ["Settings.Language.Description"] = "Use localized resources for major interface languages.",
+                ["Settings.Section.Proxy"] = "Proxy",
+                ["Settings.TransparentProxy.Title"] = "Transparent proxy",
+                ["Settings.TransparentProxy.Description"] = "Enable TUN mode; fall back to system proxy when unavailable.",
+                ["Settings.TunFallback.Title"] = "Fallback when TUN fails",
+                ["Settings.TunFallback.Description"] = "When transparent proxy cannot start, automatically take over through Windows system proxy.",
+                ["Settings.MixedPort.Title"] = "Mixed port",
+                ["Settings.MixedPort.Description"] = "Local HTTP/SOCKS entry used by system and terminal proxy settings.",
+                ["Settings.ConnectionSampling.Title"] = "Background connection sampling",
+                ["Settings.ConnectionSampling.Description"] = "Periodically read mihomo active connections and write SQLite statistics.",
+                ["Settings.SamplingInterval.Title"] = "Sampling interval",
+                ["Settings.SamplingInterval.Description"] = "Interval for background connection sampling, in seconds.",
+                ["Settings.Section.WindowsNative"] = "Native Windows",
+                ["Settings.WindowsNative.Title"] = "WSL, terminal, and Microsoft Store network repair",
+                ["Settings.WindowsNative.Description"] = "Diagnose and apply each item independently with reversible actions.",
+                ["Settings.Wsl.Title"] = "WSL network repair",
+                ["Settings.Terminal.Title"] = "Terminal network repair",
+                ["Settings.Store.Title"] = "Microsoft Store network repair",
+                ["Command.Diagnose"] = "Diagnose",
+                ["Command.Apply"] = "Apply",
+                ["Command.Reset"] = "Reset",
+                ["Diagnostic.NotRun"] = "Not diagnosed",
+                ["Settings.CheckStaleProxy.Title"] = "Check stale proxy on startup",
+                ["Settings.CheckStaleProxy.Description"] = "If Windows proxy still points to the Clash# local port, handle it with the recovery policy.",
+                ["Settings.RestoreProxyOnExit.Title"] = "Restore system proxy on exit",
+                ["Settings.RestoreProxyOnExit.Description"] = "Stop the core and disable the Windows system proxy when Clash# closes.",
+                ["Settings.ProxyRecoveryMode.Title"] = "Crash recovery policy",
+                ["Settings.ProxyRecoveryMode.Description"] = "Handles stale system proxy after another proxy app or Clash# exits unexpectedly.",
+                ["Settings.ProxyRecoveryMode.Ignore"] = "Do nothing",
+                ["Settings.ProxyRecoveryMode.Enable"] = "Restore proxy on",
+                ["Settings.ProxyRecoveryMode.Disable"] = "Restore proxy off",
+                ["Settings.Section.MainlandChina"] = "Mainland China",
+                ["Settings.MainlandChinaDisplay.Title"] = "Text and flag replacement for mainland China",
+                ["Settings.MainlandChinaDisplay.Description"] = "Enabled by default; sensitive-word filtering affects display only, not copy, search, or export.",
+            }),
+            [AppLanguage.Russian] = new(new Dictionary<string, string>
+            {
+                ["Nav.MasterControl"] = "Управление",
+                ["Nav.Proxies"] = "Прокси",
+                ["Nav.ProxyNodes"] = "Узлы",
+                ["Nav.Profiles"] = "Профили",
+                ["Nav.Links"] = "Ссылки",
+                ["Nav.Connections"] = "Соединения",
+                ["Nav.Rules"] = "Правила",
+                ["Nav.Statistics"] = "Статистика",
+                ["Nav.Logs"] = "Журналы",
+                ["Nav.Settings"] = "Настройки",
+                ["Page.Stub"] = "В разработке…",
+                ["Page.MasterControl.Description"] = "Управление состоянием Clash#, ядром и текущим прокси-входом.",
+                ["Page.ProxyNodes.Description"] = "Просмотр узлов, регионов, задержки и выбора групп прокси.",
+                ["Page.Profiles.Description"] = "Управление подписками, проверкой конфигураций, автообновлением и откатом.",
+                ["Page.Links.Description"] = "Управление ссылками подписок, периодом обновления, проверкой и импортом профилей.",
+                ["Page.Connections.Description"] = "Мониторинг соединений, процессов, совпадений правил и маршрутов прокси.",
+                ["Page.Rules.Description"] = "Просмотр наборов правил, результатов сопоставления и политики перехвата.",
+                ["Page.Statistics.Description"] = "Просмотр долгосрочной статистики по профилям, узлам и времени.",
+                ["Page.Logs.Description"] = "Просмотр журналов SQLite, занятого места и действий очистки.",
+                ["Page.Settings.Description"] = "Настройка языка, прозрачного прокси, исправлений Windows и политики отображения материкового Китая.",
+                ["Command.Add"] = "Добавить",
+                ["Command.Check"] = "Проверить",
+                ["Command.Import"] = "Импорт",
+                ["Command.PersistSnapshot"] = "Сохранить снимок",
+                ["Command.Refresh"] = "Обновить",
+                ["Command.SetActive"] = "Сделать активным",
+                ["Command.TestLatency"] = "Проверить задержку",
+                ["Command.Update"] = "Обновить",
+                ["Command.Validate"] = "Проверить",
+                ["Command.Cleanup"] = "Очистить",
+                ["Command.Cancel"] = "Отмена",
+                ["Label.CurrentProfile"] = "Текущий профиль",
+                ["Logs.Storage.Title"] = "Локальное хранилище",
+                ["Logs.StorageUsage.Format"] = "{0}, журналов: {1:N0}, соединений: {2:N0}",
+                ["Logs.Cleanup.Title"] = "Очистить локальные записи",
+                ["Logs.Cleanup.Mode.ByDate"] = "Очистить по дате",
+                ["Logs.Cleanup.Mode.BySize"] = "Очистить по размеру",
+                ["Logs.Cleanup.Mode.ByCount"] = "Очистить по количеству",
+                ["Logs.Cleanup.Mode.All"] = "Очистить все",
+                ["Logs.Cleanup.Parameter.KeepDays"] = "Дней хранить",
+                ["Logs.Cleanup.Parameter.TargetSize"] = "Целевой размер MB",
+                ["Logs.Cleanup.Parameter.KeepLogCount"] = "Журналов хранить",
+                ["Logs.Cleanup.Parameter.None"] = "Без параметра",
+                ["Logs.Cleanup.Description.ByDate"] = "Удалить журналы, соединения и снимки трафика старше выбранного числа дней.",
+                ["Logs.Cleanup.Description.BySize"] = "Удалять самые старые записи и сжимать базу данных до целевого размера.",
+                ["Logs.Cleanup.Description.ByCount"] = "Ограничить только таблицу журналов; соединения, трафик и статистика правил сохраняются.",
+                ["Logs.Cleanup.Description.All"] = "Удалить все журналы, соединения, трафик, состояние узлов и записи срабатывания правил.",
+                ["Statistics.Total.Title"] = "Итоги",
+                ["Statistics.Profile.Title"] = "Профили",
+                ["Statistics.Node.Title"] = "Узлы",
+                ["Statistics.ByProfile.Title"] = "По профилю",
+                ["Statistics.ByDate.Title"] = "По дате",
+                ["Statistics.ByNode.Title"] = "По узлу",
+                ["Statistics.LogsShortcut.Title"] = "Журналы",
+                ["Statistics.LogsShortcut.Description"] = "Просмотр журналов SQLite, занятого места и действий очистки.",
+                ["Statistics.OpenLogs"] = "Открыть журналы",
+                ["Statistics.TotalTraffic.Format"] = "Отправлено {0} / получено {1}",
+                ["Statistics.ConnectionCount.Format"] = "{0:N0} записей соединений",
+                ["Statistics.ProfileCount.Format"] = "{0:N0} агрегатов профилей",
+                ["Statistics.SnapshotCount.Format"] = "{0:N0} временных снимков",
+                ["Statistics.NodeCount.Format"] = "{0:N0} агрегатов узлов / {1:N0} записей состояния",
+                ["Statistics.RuleCount.Format"] = "{0:N0} агрегатов срабатываний правил",
+                ["Master.StatusControl.Title"] = "Основное состояние",
+                ["Master.StatusControl.Description"] = "Выберите, как Clash# перехватывает сеть Windows. Прозрачный прокси настраивается в параметрах.",
+                ["Master.Mode.Disabled.Title"] = "Отключено",
+                ["Master.Mode.Disabled.Description"] = "Остановить ядро и восстановить системный прокси.",
+                ["Master.Mode.Standby.Title"] = "Ожидание",
+                ["Master.Mode.Standby.Description"] = "Ядро работает, по умолчанию прямое подключение.",
+                ["Master.Mode.RuleTakeover.Title"] = "По правилам",
+                ["Master.Mode.RuleTakeover.Description"] = "Запустить ядро и перехватывать сеть по правилам.",
+                ["Master.Mode.FullTakeover.Title"] = "Полный перехват",
+                ["Master.Mode.FullTakeover.Description"] = "Запустить ядро и перехватывать сеть глобальной политикой.",
+                ["Master.Status.Core"] = "Состояние ядра",
+                ["Master.Status.SystemProxy"] = "Системный прокси",
+                ["Master.Status.TransparentProxy"] = "Прозрачный прокси",
+                ["Master.Status.CoreConfiguration"] = "Конфигурация ядра",
+                ["Master.Status.ConfigurationUnavailable"] = "Конфигурация недоступна",
+                ["Master.Status.CoreReady.Format"] = "Готово · {0}",
+                ["Master.Status.CoreUnavailable"] = "Ядро недоступно",
+                ["Master.Status.CoreStartFailed"] = "Не удалось запустить ядро",
+                ["Master.Status.Running"] = "Работает",
+                ["Master.Status.NotRunning"] = "Не работает",
+                ["Master.Status.On"] = "Включено",
+                ["Master.Status.Off"] = "Отключено",
+                ["Master.Status.Fallback"] = "Резервный режим",
+                ["Master.Status.Unavailable"] = "Недоступно",
+                ["Master.Status.Standby"] = "Ожидание",
+                ["Settings.Section.Language"] = "Язык",
+                ["Settings.Language.Title"] = "Язык интерфейса",
+                ["Settings.Language.Description"] = "Использовать локализованные ресурсы для основных языков интерфейса.",
+                ["Settings.Section.Proxy"] = "Прокси",
+                ["Settings.TransparentProxy.Title"] = "Прозрачный прокси",
+                ["Settings.TransparentProxy.Description"] = "Включить режим TUN; при недоступности использовать системный прокси.",
+                ["Settings.TunFallback.Title"] = "Резерв при сбое TUN",
+                ["Settings.TunFallback.Description"] = "Если прозрачный прокси не запускается, автоматически использовать системный прокси Windows.",
+                ["Settings.MixedPort.Title"] = "Смешанный порт",
+                ["Settings.MixedPort.Description"] = "Локальный вход HTTP/SOCKS для системного и терминального прокси.",
+                ["Settings.ConnectionSampling.Title"] = "Фоновая выборка соединений",
+                ["Settings.ConnectionSampling.Description"] = "Периодически читать активные соединения mihomo и писать статистику SQLite.",
+                ["Settings.SamplingInterval.Title"] = "Интервал выборки",
+                ["Settings.SamplingInterval.Description"] = "Интервал фоновой выборки соединений в секундах.",
+                ["Settings.Section.WindowsNative"] = "Windows",
+                ["Settings.WindowsNative.Title"] = "Исправление сети WSL, терминала и Microsoft Store",
+                ["Settings.WindowsNative.Description"] = "Диагностика и применение отдельно для каждого пункта с обратимыми действиями.",
+                ["Settings.Wsl.Title"] = "Исправление сети WSL",
+                ["Settings.Terminal.Title"] = "Исправление сети терминала",
+                ["Settings.Store.Title"] = "Исправление сети Microsoft Store",
+                ["Command.Diagnose"] = "Диагностика",
+                ["Command.Apply"] = "Применить",
+                ["Command.Reset"] = "Сброс",
+                ["Diagnostic.NotRun"] = "Не проверено",
+                ["Settings.CheckStaleProxy.Title"] = "Проверять остаточный прокси при запуске",
+                ["Settings.CheckStaleProxy.Description"] = "Если прокси Windows указывает на локальный порт Clash#, применить политику восстановления.",
+                ["Settings.RestoreProxyOnExit.Title"] = "Восстанавливать системный прокси при выходе",
+                ["Settings.RestoreProxyOnExit.Description"] = "Останавливать ядро и отключать системный прокси Windows при закрытии Clash#.",
+                ["Settings.ProxyRecoveryMode.Title"] = "Политика восстановления",
+                ["Settings.ProxyRecoveryMode.Description"] = "Устраняет остаточный системный прокси после сбоя Clash# или другого прокси-приложения.",
+                ["Settings.ProxyRecoveryMode.Ignore"] = "Не изменять",
+                ["Settings.ProxyRecoveryMode.Enable"] = "Восстановить прокси включенным",
+                ["Settings.ProxyRecoveryMode.Disable"] = "Восстановить прокси выключенным",
+                ["Settings.Section.MainlandChina"] = "Материковый Китай",
+                ["Settings.MainlandChinaDisplay.Title"] = "Замена текста и флагов для материкового Китая",
+                ["Settings.MainlandChinaDisplay.Description"] = "Включено по умолчанию; фильтр влияет только на отображение, не на копирование, поиск и экспорт.",
+            }),
+            [AppLanguage.French] = new(new Dictionary<string, string>
+            {
+                ["Nav.MasterControl"] = "Contrôle",
+                ["Nav.Proxies"] = "Proxies",
+                ["Nav.ProxyNodes"] = "Nœuds",
+                ["Nav.Profiles"] = "Profils",
+                ["Nav.Links"] = "Liens",
+                ["Nav.Connections"] = "Connexions",
+                ["Nav.Rules"] = "Règles",
+                ["Nav.Statistics"] = "Statistiques",
+                ["Nav.Logs"] = "Journaux",
+                ["Nav.Settings"] = "Paramètres",
+                ["Page.Stub"] = "En développement…",
+                ["Page.MasterControl.Description"] = "Contrôler l'état de prise en charge de Clash#, le noyau et l'entrée proxy actuelle.",
+                ["Page.ProxyNodes.Description"] = "Consulter les nœuds, régions, latences et choix de groupes proxy.",
+                ["Page.Profiles.Description"] = "Gérer les abonnements, la validation des configurations, les mises à jour automatiques et les retours arrière.",
+                ["Page.Links.Description"] = "Gérer les liens d'abonnement, les intervalles de mise à jour, l'état de vérification et l'import de profils.",
+                ["Page.Connections.Description"] = "Surveiller les connexions, processus, règles atteintes et routes proxy.",
+                ["Page.Rules.Description"] = "Inspecter les fournisseurs de règles, les résultats de correspondance et la politique de prise en charge.",
+                ["Page.Statistics.Description"] = "Consulter les statistiques long terme par profil, nœud et période.",
+                ["Page.Logs.Description"] = "Consulter les journaux SQLite, l'espace utilisé et les actions de nettoyage.",
+                ["Page.Settings.Description"] = "Configurer la langue, le proxy transparent, les correctifs Windows natifs et l'affichage de la Chine continentale.",
+                ["Command.Add"] = "Ajouter",
+                ["Command.Check"] = "Vérifier",
+                ["Command.Import"] = "Importer",
+                ["Command.PersistSnapshot"] = "Enregistrer",
+                ["Command.Refresh"] = "Actualiser",
+                ["Command.SetActive"] = "Activer",
+                ["Command.TestLatency"] = "Tester la latence",
+                ["Command.Update"] = "Mettre à jour",
+                ["Command.Validate"] = "Valider",
+                ["Command.Cleanup"] = "Nettoyer",
+                ["Command.Cancel"] = "Annuler",
+                ["Label.CurrentProfile"] = "Profil actuel",
+                ["Logs.Storage.Title"] = "Stockage local utilisé",
+                ["Logs.StorageUsage.Format"] = "{0}, {1:N0} journaux, {2:N0} connexions",
+                ["Logs.Cleanup.Title"] = "Nettoyer les enregistrements locaux",
+                ["Logs.Cleanup.Mode.ByDate"] = "Nettoyer par date",
+                ["Logs.Cleanup.Mode.BySize"] = "Nettoyer par taille",
+                ["Logs.Cleanup.Mode.ByCount"] = "Nettoyer par nombre",
+                ["Logs.Cleanup.Mode.All"] = "Tout nettoyer",
+                ["Logs.Cleanup.Parameter.KeepDays"] = "Jours à conserver",
+                ["Logs.Cleanup.Parameter.TargetSize"] = "Taille cible MB",
+                ["Logs.Cleanup.Parameter.KeepLogCount"] = "Journaux à conserver",
+                ["Logs.Cleanup.Parameter.None"] = "Aucun paramètre",
+                ["Logs.Cleanup.Description.ByDate"] = "Supprimer les journaux, connexions et instantanés de trafic plus anciens que le nombre de jours choisi.",
+                ["Logs.Cleanup.Description.BySize"] = "Supprimer les plus anciens enregistrements et compacter la base jusqu'à approcher la taille cible.",
+                ["Logs.Cleanup.Description.ByCount"] = "Limiter seulement le nombre de journaux ; connexions, trafic et statistiques de règles sont conservés.",
+                ["Logs.Cleanup.Description.All"] = "Supprimer tous les journaux, connexions, trafics, états de nœuds et règles atteintes.",
+                ["Statistics.Total.Title"] = "Totaux",
+                ["Statistics.Profile.Title"] = "Profils",
+                ["Statistics.Node.Title"] = "Nœuds",
+                ["Statistics.ByProfile.Title"] = "Par profil",
+                ["Statistics.ByDate.Title"] = "Par date",
+                ["Statistics.ByNode.Title"] = "Par nœud",
+                ["Statistics.LogsShortcut.Title"] = "Journaux",
+                ["Statistics.LogsShortcut.Description"] = "Consulter les journaux SQLite, l'espace utilisé et les actions de nettoyage.",
+                ["Statistics.OpenLogs"] = "Ouvrir les journaux",
+                ["Statistics.TotalTraffic.Format"] = "Envoi {0} / réception {1}",
+                ["Statistics.ConnectionCount.Format"] = "{0:N0} enregistrements de connexion",
+                ["Statistics.ProfileCount.Format"] = "{0:N0} agrégats de profils",
+                ["Statistics.SnapshotCount.Format"] = "{0:N0} instantanés temporels",
+                ["Statistics.NodeCount.Format"] = "{0:N0} agrégats de nœuds / {1:N0} états de santé",
+                ["Statistics.RuleCount.Format"] = "{0:N0} agrégats de règles atteintes",
+                ["Master.StatusControl.Title"] = "Contrôle principal",
+                ["Master.StatusControl.Description"] = "Choisissez comment Clash# prend en charge le réseau Windows. Le proxy transparent se règle dans Paramètres.",
+                ["Master.Mode.Disabled.Title"] = "Désactivé",
+                ["Master.Mode.Disabled.Description"] = "Arrêter le noyau et restaurer le proxy système.",
+                ["Master.Mode.Standby.Title"] = "Veille",
+                ["Master.Mode.Standby.Description"] = "Noyau actif, connexion directe par défaut.",
+                ["Master.Mode.RuleTakeover.Title"] = "Prise en charge par règles",
+                ["Master.Mode.RuleTakeover.Description"] = "Démarrer le noyau et prendre en charge le réseau selon les règles.",
+                ["Master.Mode.FullTakeover.Title"] = "Prise en charge globale",
+                ["Master.Mode.FullTakeover.Description"] = "Démarrer le noyau et prendre en charge le réseau en mode global.",
+                ["Master.Status.Core"] = "État du noyau",
+                ["Master.Status.SystemProxy"] = "Proxy système",
+                ["Master.Status.TransparentProxy"] = "Proxy transparent",
+                ["Master.Status.CoreConfiguration"] = "Configuration du noyau",
+                ["Master.Status.ConfigurationUnavailable"] = "Configuration indisponible",
+                ["Master.Status.CoreReady.Format"] = "Prêt · {0}",
+                ["Master.Status.CoreUnavailable"] = "Noyau indisponible",
+                ["Master.Status.CoreStartFailed"] = "Échec du démarrage du noyau",
+                ["Master.Status.Running"] = "En cours",
+                ["Master.Status.NotRunning"] = "Arrêté",
+                ["Master.Status.On"] = "Activé",
+                ["Master.Status.Off"] = "Désactivé",
+                ["Master.Status.Fallback"] = "Repli actif",
+                ["Master.Status.Unavailable"] = "Indisponible",
+                ["Master.Status.Standby"] = "Veille",
+                ["Settings.Section.Language"] = "Langue",
+                ["Settings.Language.Title"] = "Langue d'affichage",
+                ["Settings.Language.Description"] = "Utiliser les ressources localisées pour les principales langues d'interface.",
+                ["Settings.Section.Proxy"] = "Proxy",
+                ["Settings.TransparentProxy.Title"] = "Proxy transparent",
+                ["Settings.TransparentProxy.Description"] = "Activer le mode TUN ; revenir au proxy système si nécessaire.",
+                ["Settings.TunFallback.Title"] = "Repli si TUN échoue",
+                ["Settings.TunFallback.Description"] = "Si le proxy transparent ne démarre pas, prendre automatiquement en charge via le proxy système Windows.",
+                ["Settings.MixedPort.Title"] = "Port mixte",
+                ["Settings.MixedPort.Description"] = "Entrée HTTP/SOCKS locale utilisée par le proxy système et le terminal.",
+                ["Settings.ConnectionSampling.Title"] = "Échantillonnage en arrière-plan",
+                ["Settings.ConnectionSampling.Description"] = "Lire périodiquement les connexions mihomo actives et écrire les statistiques SQLite.",
+                ["Settings.SamplingInterval.Title"] = "Intervalle d'échantillonnage",
+                ["Settings.SamplingInterval.Description"] = "Intervalle d'échantillonnage en arrière-plan, en secondes.",
+                ["Settings.Section.WindowsNative"] = "Windows natif",
+                ["Settings.WindowsNative.Title"] = "Réparation réseau WSL, terminal et Microsoft Store",
+                ["Settings.WindowsNative.Description"] = "Diagnostiquer et appliquer chaque élément séparément avec actions réversibles.",
+                ["Settings.Wsl.Title"] = "Réparation réseau WSL",
+                ["Settings.Terminal.Title"] = "Réparation réseau du terminal",
+                ["Settings.Store.Title"] = "Réparation réseau Microsoft Store",
+                ["Command.Diagnose"] = "Diagnostiquer",
+                ["Command.Apply"] = "Appliquer",
+                ["Command.Reset"] = "Réinitialiser",
+                ["Diagnostic.NotRun"] = "Non diagnostiqué",
+                ["Settings.CheckStaleProxy.Title"] = "Vérifier le proxy résiduel au démarrage",
+                ["Settings.CheckStaleProxy.Description"] = "Si le proxy Windows pointe encore vers le port local Clash#, appliquer la politique de récupération.",
+                ["Settings.RestoreProxyOnExit.Title"] = "Restaurer le proxy système à la fermeture",
+                ["Settings.RestoreProxyOnExit.Description"] = "Arrêter le noyau et désactiver le proxy système Windows à la fermeture de Clash#.",
+                ["Settings.ProxyRecoveryMode.Title"] = "Politique de récupération",
+                ["Settings.ProxyRecoveryMode.Description"] = "Gère le proxy système résiduel après fermeture anormale de Clash# ou d'une autre application proxy.",
+                ["Settings.ProxyRecoveryMode.Ignore"] = "Ne rien faire",
+                ["Settings.ProxyRecoveryMode.Enable"] = "Restaurer proxy activé",
+                ["Settings.ProxyRecoveryMode.Disable"] = "Restaurer proxy désactivé",
+                ["Settings.Section.MainlandChina"] = "Chine continentale",
+                ["Settings.MainlandChinaDisplay.Title"] = "Remplacement du texte et des drapeaux pour la Chine continentale",
+                ["Settings.MainlandChinaDisplay.Description"] = "Activé par défaut ; le filtrage affecte seulement l'affichage, pas la copie, la recherche ni l'export.",
+            }),
+            [AppLanguage.German] = new(new Dictionary<string, string>
+            {
+                ["Nav.MasterControl"] = "Steuerung",
+                ["Nav.Proxies"] = "Proxies",
+                ["Nav.ProxyNodes"] = "Knoten",
+                ["Nav.Profiles"] = "Profile",
+                ["Nav.Links"] = "Links",
+                ["Nav.Connections"] = "Verbindungen",
+                ["Nav.Rules"] = "Regeln",
+                ["Nav.Statistics"] = "Statistiken",
+                ["Nav.Logs"] = "Protokolle",
+                ["Nav.Settings"] = "Einstellungen",
+                ["Page.Stub"] = "In Entwicklung…",
+                ["Page.MasterControl.Description"] = "Clash#-Übernahmestatus, Kernlaufzeit und aktuellen Proxy-Einstieg steuern.",
+                ["Page.ProxyNodes.Description"] = "Knoten, Regionen, Latenz und aktive Proxygruppen-Auswahl prüfen.",
+                ["Page.Profiles.Description"] = "Abonnements, Konfigurationsprüfung, automatische Updates und Rollback verwalten.",
+                ["Page.Links.Description"] = "Abonnementlinks, Updateintervalle, Prüfstatus und Profilimport verwalten.",
+                ["Page.Connections.Description"] = "Aktive Verbindungen, Prozesse, Regel-Treffer und Proxy-Routen überwachen.",
+                ["Page.Rules.Description"] = "Regelanbieter, Treffergebnisse und Übernahmerichtlinie prüfen.",
+                ["Page.Statistics.Description"] = "Langzeitstatistiken nach Profil, Knoten und Zeitraum prüfen.",
+                ["Page.Logs.Description"] = "SQLite-Protokolle, Speicherbelegung und Bereinigungsaktionen prüfen.",
+                ["Page.Settings.Description"] = "Sprache, transparenten Proxy, native Windows-Reparaturen und Anzeige für Festlandchina konfigurieren.",
+                ["Command.Add"] = "Hinzufügen",
+                ["Command.Check"] = "Prüfen",
+                ["Command.Import"] = "Importieren",
+                ["Command.PersistSnapshot"] = "Momentaufnahme",
+                ["Command.Refresh"] = "Aktualisieren",
+                ["Command.SetActive"] = "Aktiv setzen",
+                ["Command.TestLatency"] = "Latenz testen",
+                ["Command.Update"] = "Aktualisieren",
+                ["Command.Validate"] = "Validieren",
+                ["Command.Cleanup"] = "Bereinigen",
+                ["Command.Cancel"] = "Abbrechen",
+                ["Label.CurrentProfile"] = "Aktuelles Profil",
+                ["Logs.Storage.Title"] = "Lokale Speicherbelegung",
+                ["Logs.StorageUsage.Format"] = "{0}, {1:N0} Protokolle, {2:N0} Verbindungsdatensätze",
+                ["Logs.Cleanup.Title"] = "Lokale Datensätze bereinigen",
+                ["Logs.Cleanup.Mode.ByDate"] = "Nach Datum bereinigen",
+                ["Logs.Cleanup.Mode.BySize"] = "Nach Größe bereinigen",
+                ["Logs.Cleanup.Mode.ByCount"] = "Nach Anzahl bereinigen",
+                ["Logs.Cleanup.Mode.All"] = "Alles bereinigen",
+                ["Logs.Cleanup.Parameter.KeepDays"] = "Tage behalten",
+                ["Logs.Cleanup.Parameter.TargetSize"] = "Zielgröße MB",
+                ["Logs.Cleanup.Parameter.KeepLogCount"] = "Protokolle behalten",
+                ["Logs.Cleanup.Parameter.None"] = "Kein Parameter",
+                ["Logs.Cleanup.Description.ByDate"] = "Protokolle, Verbindungen und Traffic-Momentaufnahmen löschen, die älter als die gewählte Tageszahl sind.",
+                ["Logs.Cleanup.Description.BySize"] = "Älteste Datensätze löschen und die Datenbank bis nahe an die Zielgröße komprimieren.",
+                ["Logs.Cleanup.Description.ByCount"] = "Nur die Anzahl der Protokolle begrenzen; Verbindungen, Traffic und Regelstatistiken bleiben erhalten.",
+                ["Logs.Cleanup.Description.All"] = "Alle Protokolle, Verbindungen, Traffic-, Knotenzustands- und Regel-Treffer-Datensätze löschen.",
+                ["Statistics.Total.Title"] = "Gesamt",
+                ["Statistics.Profile.Title"] = "Profile",
+                ["Statistics.Node.Title"] = "Knoten",
+                ["Statistics.ByProfile.Title"] = "Nach Profil",
+                ["Statistics.ByDate.Title"] = "Nach Datum",
+                ["Statistics.ByNode.Title"] = "Nach Knoten",
+                ["Statistics.LogsShortcut.Title"] = "Protokollsystem",
+                ["Statistics.LogsShortcut.Description"] = "SQLite-Protokolle, Speicherbelegung und Bereinigungsaktionen prüfen.",
+                ["Statistics.OpenLogs"] = "Protokolle öffnen",
+                ["Statistics.TotalTraffic.Format"] = "Upload {0} / Download {1}",
+                ["Statistics.ConnectionCount.Format"] = "{0:N0} Verbindungsdatensätze",
+                ["Statistics.ProfileCount.Format"] = "{0:N0} Profilaggregate",
+                ["Statistics.SnapshotCount.Format"] = "{0:N0} Zeitmomentaufnahmen",
+                ["Statistics.NodeCount.Format"] = "{0:N0} Knotenaggregate / {1:N0} Zustandsdatensätze",
+                ["Statistics.RuleCount.Format"] = "{0:N0} Regel-Trefferaggregate",
+                ["Master.StatusControl.Title"] = "Primäre Statussteuerung",
+                ["Master.StatusControl.Description"] = "Wählen Sie, wie Clash# das Windows-Netzwerk übernimmt. Transparenter Proxy wird in den Einstellungen konfiguriert.",
+                ["Master.Mode.Disabled.Title"] = "Deaktiviert",
+                ["Master.Mode.Disabled.Description"] = "Kern stoppen und Systemproxy wiederherstellen.",
+                ["Master.Mode.Standby.Title"] = "Bereitschaft",
+                ["Master.Mode.Standby.Description"] = "Kern läuft, standardmäßig direkte Verbindung.",
+                ["Master.Mode.RuleTakeover.Title"] = "Regelübernahme",
+                ["Master.Mode.RuleTakeover.Description"] = "Kern starten und Netzwerk per Regelrichtlinie übernehmen.",
+                ["Master.Mode.FullTakeover.Title"] = "Vollständige Übernahme",
+                ["Master.Mode.FullTakeover.Description"] = "Kern starten und Netzwerk per globaler Richtlinie übernehmen.",
+                ["Master.Status.Core"] = "Kernstatus",
+                ["Master.Status.SystemProxy"] = "Systemproxy",
+                ["Master.Status.TransparentProxy"] = "Transparenter Proxy",
+                ["Master.Status.CoreConfiguration"] = "Kernkonfiguration",
+                ["Master.Status.ConfigurationUnavailable"] = "Konfiguration nicht verfügbar",
+                ["Master.Status.CoreReady.Format"] = "Bereit · {0}",
+                ["Master.Status.CoreUnavailable"] = "Kern nicht verfügbar",
+                ["Master.Status.CoreStartFailed"] = "Kernstart fehlgeschlagen",
+                ["Master.Status.Running"] = "Wird ausgeführt",
+                ["Master.Status.NotRunning"] = "Nicht ausgeführt",
+                ["Master.Status.On"] = "Ein",
+                ["Master.Status.Off"] = "Aus",
+                ["Master.Status.Fallback"] = "Fallback aktiv",
+                ["Master.Status.Unavailable"] = "Nicht verfügbar",
+                ["Master.Status.Standby"] = "Bereitschaft",
+                ["Settings.Section.Language"] = "Sprache",
+                ["Settings.Language.Title"] = "Anzeigesprache",
+                ["Settings.Language.Description"] = "Lokalisierte Ressourcen für wichtige Oberflächensprachen verwenden.",
+                ["Settings.Section.Proxy"] = "Proxy",
+                ["Settings.TransparentProxy.Title"] = "Transparenter Proxy",
+                ["Settings.TransparentProxy.Description"] = "TUN-Modus aktivieren; bei Fehlern auf Systemproxy zurückfallen.",
+                ["Settings.TunFallback.Title"] = "Fallback bei TUN-Fehler",
+                ["Settings.TunFallback.Description"] = "Wenn der transparente Proxy nicht startet, automatisch per Windows-Systemproxy übernehmen.",
+                ["Settings.MixedPort.Title"] = "Gemischter Port",
+                ["Settings.MixedPort.Description"] = "Lokaler HTTP/SOCKS-Einstieg für System- und Terminalproxy.",
+                ["Settings.ConnectionSampling.Title"] = "Verbindungsabtastung im Hintergrund",
+                ["Settings.ConnectionSampling.Description"] = "Aktive mihomo-Verbindungen regelmäßig lesen und SQLite-Statistiken schreiben.",
+                ["Settings.SamplingInterval.Title"] = "Abtastintervall",
+                ["Settings.SamplingInterval.Description"] = "Intervall der Hintergrundabtastung in Sekunden.",
+                ["Settings.Section.WindowsNative"] = "Windows-nativ",
+                ["Settings.WindowsNative.Title"] = "Netzwerkreparatur für WSL, Terminal und Microsoft Store",
+                ["Settings.WindowsNative.Description"] = "Jeden Punkt separat diagnostizieren und anwenden, mit rückgängig machbaren Aktionen.",
+                ["Settings.Wsl.Title"] = "WSL-Netzwerkreparatur",
+                ["Settings.Terminal.Title"] = "Terminal-Netzwerkreparatur",
+                ["Settings.Store.Title"] = "Microsoft Store-Netzwerkreparatur",
+                ["Command.Diagnose"] = "Diagnose",
+                ["Command.Apply"] = "Anwenden",
+                ["Command.Reset"] = "Zurücksetzen",
+                ["Diagnostic.NotRun"] = "Nicht diagnostiziert",
+                ["Settings.CheckStaleProxy.Title"] = "Veralteten Proxy beim Start prüfen",
+                ["Settings.CheckStaleProxy.Description"] = "Wenn der Windows-Proxy noch auf den lokalen Clash#-Port zeigt, die Wiederherstellungsrichtlinie anwenden.",
+                ["Settings.RestoreProxyOnExit.Title"] = "Systemproxy beim Beenden wiederherstellen",
+                ["Settings.RestoreProxyOnExit.Description"] = "Beim Schließen von Clash# den Kern stoppen und den Windows-Systemproxy deaktivieren.",
+                ["Settings.ProxyRecoveryMode.Title"] = "Wiederherstellungsrichtlinie",
+                ["Settings.ProxyRecoveryMode.Description"] = "Behandelt veraltete Systemproxy-Einstellungen nach unerwartetem Beenden von Clash# oder anderen Proxy-Apps.",
+                ["Settings.ProxyRecoveryMode.Ignore"] = "Nichts tun",
+                ["Settings.ProxyRecoveryMode.Enable"] = "Proxy eingeschaltet wiederherstellen",
+                ["Settings.ProxyRecoveryMode.Disable"] = "Proxy ausgeschaltet wiederherstellen",
+                ["Settings.Section.MainlandChina"] = "Festlandchina",
+                ["Settings.MainlandChinaDisplay.Title"] = "Text- und Flaggenersetzung für Festlandchina",
+                ["Settings.MainlandChinaDisplay.Description"] = "Standardmäßig aktiviert; Filterung betrifft nur Anzeige, nicht Kopieren, Suche oder Export.",
+            }),
+        };
+
+        return new ReadOnlyDictionary<AppLanguage, ReadOnlyDictionary<string, string>>(table);
+    }
+}
