+++ ClashSharp/ClashSharp/Themes/Generic.xaml
@@ -1,5 +1,19 @@
 <ResourceDictionary
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:controls="using:Microsoft.UI.Xaml.Controls">
 
+    <ResourceDictionary.MergedDictionaries>
+        <controls:XamlControlsResources />
+    </ResourceDictionary.MergedDictionaries>
+
+    <Thickness x:Key="ClashPagePadding">32</Thickness>
+
+    <Style x:Key="ClashCardGridStyle" TargetType="Grid">
+        <Setter Property="Background" Value="{ThemeResource CardBackgroundFillColorDefaultBrush}" />
+        <Setter Property="BorderBrush" Value="{ThemeResource CardStrokeColorDefaultBrush}" />
+        <Setter Property="BorderThickness" Value="1" />
+        <Setter Property="CornerRadius" Value="4" />
+        <Setter Property="Padding" Value="16" />
+    </Style>
 </ResourceDictionary>
